@@ -3,10 +3,15 @@
   "version": "0.2.4",
   "description": "一款PC端视频聚合应用，提供国内主流视频站资源。",
   "main": "app/main.js",
+  "packageManager": "yarn@3.2.3",
   "scripts": {
+    "electron": "electron .",
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "watch": "tsc --watch",
+    "forge-build": "cd ./ts && webpack",
+    "preview": "vite preview",
     "start": "electron-forge start",
-    "watch": "cd ./ts && webpack --watch",
-    "build": "cd ./ts && webpack",
     "test": "npm run model-test",
     "model-test": "tsc -p . && electron-mocha --renderer built/test/model --enable-logging",
     "ui-test": "tsc -p . && nyc mocha built/test/ui && nyc report --reporter=json && codecov -f coverage/*.json",
@@ -34,50 +39,36 @@
     "url": "https://github.com/sunxfancy"
   },
   "license": "MIT",
-  "devDependencies": {
-    "@types/chai": "^3.4.34",
-    "@types/chai-as-promised": "0.0.29",
-    "@types/electron": "^1.4.31",
-    "@types/material-ui": "^0.16.48",
-    "@types/mocha": "^2.2.38",
-    "@types/nedb": "0.0.31",
-    "@types/react": "^15.0.5",
-    "@types/react-dom": "^0.14.22",
-    "@types/react-router": "^3.0.0",
-    "@types/request": "0.0.39",
-    "@types/webpack-env": "^1.13.0",
-    "chai": "^3.5.0",
-    "chai-as-promised": "^6.0.0",
-    "codecov": "^1.0.1",
-    "devtron": "^1.3.0",
-    "dts-generator": "^2.0.0",
-    "electron-forge": "^2.12.0",
-    "electron-mocha": "^3.3.0",
-    "electron-prebuilt-compile": "1.6.6",
-    "electron-react-devtools": "^0.4.0",
-    "github-download": "^0.5.0",
-    "material-ui": "^0.16.7",
-    "mocha": "^3.2.0",
-    "nedb": "^1.8.0",
-    "nodeify": "^1.0.1",
-    "nyc": "^10.1.0",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4",
-    "react-html5video": "^1.4.0",
-    "react-layout-components": "^3.0.1",
-    "react-materialui-notifications": "^0.4.1",
-    "react-router": "^3.0.2",
-    "react-swipeable-views": "^0.10.3",
-    "react-swipeable-views-utils": "^0.10.1",
-    "react-tap-event-plugin": "^2.0.1",
-    "request": "^2.79.0",
-    "spectron": "~3.4.1",
-    "ts-loader": "^2.0.3",
-    "typescript": "^2.3.2",
-    "webpack": "^2.5.1",
-    "youku-client": "^0.1.2"
-  },
   "config": {
     "forge": "./forge.config.js"
+  },
+  "dependencies": {
+    "@electron/remote": "^2.0.9",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/material": "^5.11.4",
+    "@types/nedb": "^1.8.12",
+    "nedb": "^1.8.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-html5video": "^2.5.1",
+    "react-layout-components": "^3.0.6",
+    "react-router": "^6.6.2",
+    "react-router-dom": "^6.6.2",
+    "request": "^2.88.2"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.20.12",
+    "@swc/cli": "^0.1.59",
+    "@swc/core": "^1.3.26",
+    "@vitejs/plugin-react-swc": "^3.0.1",
+    "electron": "^22.0.2",
+    "electron-forge": "^5.2.4",
+    "ts-loader": "^9.4.2",
+    "typescript": "^4.9.4",
+    "vite": "^4.0.4",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1"
   }
 }
