@@ -1,10 +1,6 @@
-import React = require('react')
-import { List, ListItem, Divider, Drawer, Paper, FlatButton } from 'material-ui'
-import { NavigationFullscreen, ImagePhotoCamera, NavigationFullscreenExit, ImageBrightness2, NotificationSync, 
-    NavigationChevronLeft, NavigationChevronRight } from 'material-ui/svg-icons';
-const {Box, VBox, Page, Container} = require('react-layout-components')
+import * as React from 'react'
 import {VideoCollection} from '../Model/resource'
-import {Link} from 'react-router'
+import {Link} from 'react-router-dom'
 
 
 export interface ShowBoxProps {
