@@ -1,8 +1,4 @@
-import React = require('react')
-import { List, ListItem, Divider, Drawer, Paper, FlatButton } from 'material-ui'
-import { NavigationFullscreen, ImagePhotoCamera, NavigationFullscreenExit, ImageBrightness2, NotificationSync, 
-    NavigationChevronLeft, NavigationChevronRight } from 'material-ui/svg-icons';
-const {Box, VBox, Page, Container} = require('react-layout-components')
+import * as React from 'react'
 import {Video} from '../Model/resource'
 
 export interface VideoBoxProps {
