@@ -1,5 +1,6 @@
-import React = require('react')
-import {Tabs, Tab, AppBar, Card, CardHeader, CardText} from 'material-ui'
+import * as React from 'react'
+import {Button, Card, CardActions, CardHeader, CardMedia, CardContent} from '@mui/material'
+
 import {Categories, DiscoveryPage} from '.'
 const {Box, VBox, Page, Container} = require('react-layout-components')
 
@@ -14,9 +15,7 @@ export class DownloadPage extends React.Component<any, any> {
             <div className='content-card'>
                 <Card>
                     <CardHeader title="下载管理"/>
-                    <CardText>
-                        
-                    </CardText>
+                    
                 </Card> 
             </div>
         </div>
