@@ -1,9 +1,9 @@
-import React = require('react')
-import {Card, CardHeader, CardText,FlatButton,TextField,SelectField,MenuItem} from 'material-ui'
+import * as React from 'react'
+import {Card, CardHeader, CardContent,Button,TextField,Select,MenuItem, RadioGroup} from '@mui/material'
 import {Categories, DiscoveryPage} from '.'
 const {Box, VBox, Page, Container} = require('react-layout-components')
-import {remote} from 'electron';
-import {App, Plugin} from '../Model'
+import {BrowserWindow} from '@electron/remote';
+import {App} from '../Model'
 
 export class SettingsPage extends React.Component<any, any> {
     constructor(props) {
@@ -12,7 +12,6 @@ export class SettingsPage extends React.Component<any, any> {
     }
 
     openInNewWindow(path) {
-        const BrowserWindow = remote.BrowserWindow;
         var win = new BrowserWindow({ width: 800, height: 600 });
         win.loadURL(path);
     }
@@ -22,20 +21,20 @@ export class SettingsPage extends React.Component<any, any> {
     }
 
     renderPlugins() {
-        let ps = App.getPlugins().plugins
-        console.log('ps', ps)
-        let ret = []
-        for (let i in ps) {
-            let p:Plugin = ps[i]
-            if (p!=null)
-            ret.push(
-                <div key={i}> 
-                    <h2>{i}</h2>
-                    <h3>{p.version}</h3>
-                </div>
-            )
-        }
-        return ret
+        // let ps = App.getPlugins().plugins
+        // console.log('ps', ps)
+        // let ret = []
+        // for (let i in ps) {
+        //     let p:Plugin = ps[i]
+        //     if (p!=null)
+        //     ret.push(
+        //         <div key={i}> 
+        //             <h2>{i}</h2>
+        //             <h3>{p.version}</h3>
+        //         </div>
+        //     )
+        // }
+        // return ret
     }
 
     render() {
@@ -46,32 +45,32 @@ export class SettingsPage extends React.Component<any, any> {
                         title="应用基本设置"
                         actAsExpander={true}
                         showExpandableButton={true} />
-                    <CardText>
+                    <CardContent>
                         <VBox>
                         <TextField
                             defaultValue={this.getDefaultPath()}
-                            floatingLabelText="视频下载路径"
+                            label="视频下载路径"
                         />
-                        <SelectField 
-                            floatingLabelText="优先选择清晰度"
+                        <Select 
+                            label="优先选择清晰度"
                             value={this.state.qxd}
                             onChange={((e, i, v)=> this.setState({qxd: v})).bind(this)}
                         >
-                            <MenuItem value={1} primaryText="标清" />
-                            <MenuItem value={2} primaryText="高清" />
-                            <MenuItem value={3} primaryText="超清" />
-                        </SelectField>
+                            <MenuItem value={1}>标清</MenuItem>
+                            <MenuItem value={2}>高清</MenuItem>
+                            <MenuItem value={3}>超清</MenuItem>
+                        </Select>
                         </VBox>
-                    </CardText>
-                    <CardText expandable={true}>
+                    </CardContent>
+                    <CardContent>
                         <VBox>
                         <h4>Chrome Info</h4>
                         <Box>
-                            <FlatButton onClick={()=>this.openInNewWindow('chrome://gpu')}>GPU info</FlatButton>
-                            <FlatButton onClick={()=>this.openInNewWindow('chrome://media-internals/')}>Media</FlatButton>
+                            <Button onClick={()=>this.openInNewWindow('chrome://gpu')}>GPU info</Button>
+                            <Button onClick={()=>this.openInNewWindow('chrome://media-internals/')}>Media</Button>
                         </Box>
                         </VBox>
-                    </CardText>
+                    </CardContent>
                 </Card> 
             </div>
             <div className='content-card'>
@@ -80,23 +79,22 @@ export class SettingsPage extends React.Component<any, any> {
                         title="扩展程序设置" 
                         actAsExpander={true}
                         showExpandableButton={true} />
-                    <CardText>
+                    <CardContent>
                         <VBox>
                             {this.renderPlugins()}
                         </VBox>
-                    </CardText>
-                    <CardText expandable={true}>
+                    </CardContent>
+                    <CardContent>
                         
-                    </CardText>
+                    </CardContent>
                 </Card>
             </div>
             <div className='content-card'>
                 <Card>
-                    <CardText>
+                    <CardContent>
                         <h1>Video Client 多功能视频客户端</h1>
                         <h3>软件版本: v{process.env.npm_package_version}</h3>
-                        <h3>作者: 西风逍遥游 (sunxfancy@gmail.com)</h3>
-                    </CardText>
+                    </CardContent>
                 </Card>
             </div>
         </div>
