@@ -1,12 +1,9 @@
-
-const remote = require("electron").remote;
+import * as remote from '@electron/remote'
 const cp = remote.require('child_process');
-const app = remote.app;
-
 export class Cmd {
     public static async run(shell:string, args:string[]): Promise<string> {
         return new Promise<string>((resolve, reject) => {
-            cp.execFile(shell, args, {env: remote.process.env},
+            cp.execFile(shell, args, {env: process.env},
                 (error, stdout, stderr) => {
                     if (error != null) {
                         if (stderr != null)
@@ -22,7 +19,7 @@ export class Cmd {
     // 令electron作为node执行新进程
     public static async run_as_node(js:string, args:string[]=null): Promise<string> {
         return new Promise<string>((resolve, reject)=>{
-            cp.fork(js, args, {env: remote.process.env},
+            cp.fork(js, args, {env: process.env},
                 (error, stdout, stderr)=> {
                     if (error != null) {
                         console.log(error)
