@@ -1,29 +1,27 @@
 import * as React from 'react'
-import { List, ListItem, Divider, Drawer } from 'material-ui'
-import { ActionHome, ActionGrade, ActionHistory, FileCloudDownload, ActionSettings, 
-    NavigationChevronLeft, NavigationChevronRight } from 'material-ui/svg-icons';
+import { List, ListItem,ListItemIcon, Divider } from '@mui/material'
+import { Home, Grade, History, CloudDownload, Settings, ChevronLeft, ChevronRight } from '@mui/icons-material';
 
 export class LeftMenu extends React.Component<any, any> {
-    static closeIcon = <NavigationChevronRight />
-    static openIcon = <NavigationChevronLeft />
+    static closeIcon = <ChevronRight />
+    static openIcon = <ChevronLeft />
     render() {
         let ss = {}
         if (!this.props.open)
             ss = {marginLeft: this.props.width+10, width: 64, overflow:'hidden'}
         return <div style={ss}>
             <div style={{width: this.props.width}}>
-                <List>
-                    <ListItem primaryText="主页" href='#/home' leftIcon={<ActionHome />} />
-                    <ListItem primaryText="我的收藏" href='#/favorite' leftIcon={<ActionGrade />} />
-                    <ListItem primaryText="历史观看" href='#/history' leftIcon={<ActionHistory />} />
-                    <ListItem primaryText="离线视频" href='#/download' leftIcon={<FileCloudDownload />} />
+                {/* <List>
+                    <ListItem href='#/home'><ListItemIcon><Home /></ListItemIcon>主页</ListItem>
+                    <ListItem href='#/favorite'><ListItemIcon><Grade /></ListItemIcon>我的收藏</ListItem>
+                    <ListItem href='#/history'><ListItemIcon><History /></ListItemIcon>历史观看</ListItem>
+                    <ListItem href='#/download'><ListItemIcon><CloudDownload /></ListItemIcon>离线视频</ListItem>
                 </List>
                 <Divider />
                 <List>
-                    <ListItem primaryText="设置" href='#/settings' leftIcon={<ActionSettings />} />
-                    <ListItem primaryText="隐藏侧边栏" onTouchTap={this.props.onBack} 
-                        leftIcon={this.props.open? LeftMenu.openIcon: LeftMenu.closeIcon} />
-                </List>
+                    <ListItem href='#/settings'><ListItemIcon><Settings /></ListItemIcon>设置</ListItem>
+                    <ListItem onClick={this.props.onBack}><ListItemIcon>{this.props.open? LeftMenu.openIcon: LeftMenu.closeIcon}</ListItemIcon>隐藏侧边栏</ListItem>
+                </List> */}
             </div>
         </div>
             
