@@ -1,6 +1,6 @@
 export * from './app'
 export * from './config'
-export * from './plugins'
+// export * from './plugins'
 export * from './resource'
 export * from './res-adapter'
 export * from './ui'
